Picture link styling + description type added
Moved the picture links below the animal's picture (as opposed to to-the-side)

Also added a 'description' type, which removes any content between ()s or []s. This is done so that the initial's of the writer of the desciption are removed before it is displayed on the page. This creates a much nicer looking desciption.