Added BE color to Pull Animals page
Added the Behavior Evaluation color to the Pull Animals page in the form of a colored circle next to the animal's name. Hovering over this circle shows a description of what the colors mean.

Added jquery-ui CSS to the pull_animals page so that the tooltip hover feature works and is styled correctly.

Also added an is_numeric check for viewanimal.php, to prevent any funky data from getting through.