Added format_name, cleaned up format_breed
Added a format_name function, which chops off the animal's name if it is over 18 characters, and adds ".." to the end to indicate the name has been chopped.

Cleaned up format_breed to prevent a redudant if/else statement that was used to chop the breed string if it was over 22 characters long.