Changed sort buttons toggling
Changed sort buttons so that, when clicking between them, it resets to the default ascending order, instead of remembering the previous order, and simply switching that.

For instance, in the old system if you were sorting by name A-Z, and then you switched to Breed A-Z, then back to Name, the sorting would be in Z-A order. Now, when you switch back, it will still be A-Z.

Clicking the same button twice in a row switches the order, the same as how it did previously. This change only affects switching between different buttons.