Made animal name page title dynamic
Made the animal's name page title change dynamic, so that it takes the current title of the page instead of a static "View Animal | Sacramento SPCA". This allows for other sites using the plugin to have their own title instead of the Sacramento SPCA's.

Also cleaned up comment indentation, and some function indentation.