Changed config pulls to async
Changed the pulling of config.json from a sync to an async function. This fixed a small halt whenever sorting buttons for Dogs were clicked.

Turned plugin_base into a global variable so it doesn't have to be passed around so much.

Also changed the styling of animal picture buttons. They are now circles and centered under the animal's picture.